Fix pressure type

+++ nettigo_air_monitor/model.py
@@ -10,7 +10,7 @@ class NAMSensors:
     bme280_humidity: Optional[float]
     bme280_pressure: Optional[int]
     bme280_temperature: Optional[float]
-    bmp180_pressure: Optional[int]
+    bmp180_pressure: Optional[float]
     bmp180_temperature: Optional[float]
     bmp280_pressure: Optional[int]
     bmp280_temperature: Optional[float]
